Добавлены тесты по заданию: - Тест на корректно сформированный запрос - Тест на возврат кода ответа 400 при запросе не поддерживаемого города - Тест на вывод всех доступных кафе, если в запросе указано большее количество чем их есть
