+++ .gitignore
@@ -1,0 +1,3 @@
+
+# задание
+precode.go
